New folder/requirements.txt
fastapi
uvicorn
nvidia-riva-client
